exam_checker/requirements.txt
openai>=1.35.0
torch>=2.3.0
transformers>=4.42.0
sentence-transformers>=3.0.0
pix2text>=1.0.1
DECIMER>=2.3.2
easyocr>=1.7.1
pytesseract>=0.3.10
PyMuPDF>=1.24.5
Pillow>=10.3.0
opencv-python>=4.10.0
sympy>=1.12
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
pandas>=2.2.2
openpyxl>=3.1.4
python-dotenv>=1.0.1
jinja2>=3.1.4
aiofiles>=23.2.1
RestrictedPython>=7.1
sqlparse>=0.5.0
numpy>=1.26.0
scikit-learn>=1.5.0
tqdm>=4.66.4
